+++ source/Appccelerate.StateMachine.Specs/Appccelerate.StateMachine.Specs.csproj
@@ -72,32 +72,32 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ActiveStateMachines.cs" />
+    <Compile Include="Sync\ActiveStateMachines.cs" />
     <Compile Include="Async\CurrentStateExtension.cs" />
     <Compile Include="Async\EntryActions.cs" />
     <Compile Include="Async\ExitActions.cs" />
     <Compile Include="Async\Guards.cs" />
     <Compile Include="Async\Transitions.cs" />
     <Compile Include="Catch.cs" />
-    <Compile Include="CurrentStateExtension.cs" />
-    <Compile Include="CustomTypes.cs" />
-    <Compile Include="EntryActions.cs" />
-    <Compile Include="ExceptionHandling.cs" />
-    <Compile Include="ExitActions.cs" />
+    <Compile Include="Sync\CurrentStateExtension.cs" />
+    <Compile Include="Sync\CustomTypes.cs" />
+    <Compile Include="Sync\EntryActions.cs" />
+    <Compile Include="Sync\ExceptionHandling.cs" />
+    <Compile Include="Sync\ExitActions.cs" />
     <Compile Include="Properties\GlobalAssemblyInfo.cs" />
-    <Compile Include="StateMachineExtensions.cs" />
-    <Compile Include="Guards.cs" />
-    <Compile Include="HierarchicalStateMachineInitialization.cs" />
-    <Compile Include="HierarchicalTransitions.cs" />
-    <Compile Include="Initialization.cs" />
-    <Compile Include="PassiveStateMachines.cs" />
-    <Compile Include="Persisting.cs" />
+    <Compile Include="Sync\StateMachineExtensions.cs" />
+    <Compile Include="Sync\Guards.cs" />
+    <Compile Include="Sync\HierarchicalStateMachineInitialization.cs" />
+    <Compile Include="Sync\HierarchicalTransitions.cs" />
+    <Compile Include="Sync\Initialization.cs" />
+    <Compile Include="Sync\PassiveStateMachines.cs" />
+    <Compile Include="Sync\Persisting.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RecordEventsExtension.cs" />
-    <Compile Include="Reporting.cs" />
-    <Compile Include="StartStop.cs" />
-    <Compile Include="StateMachineNameReporter.cs" />
-    <Compile Include="Transitions.cs" />
+    <Compile Include="Sync\RecordEventsExtension.cs" />
+    <Compile Include="Sync\Reporting.cs" />
+    <Compile Include="Sync\StartStop.cs" />
+    <Compile Include="Sync\StateMachineNameReporter.cs" />
+    <Compile Include="Sync\Transitions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Analyzers.ruleset" />
@@ -105,7 +105,6 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
-    <None Include="Settings.stylecop" />
     <AdditionalFiles Include="stylecop.json" />
   </ItemGroup>
   <ItemGroup>
