+++ source/Appccelerate.StateMachine.Specs/Sync/ExceptionHandling.cs
@@ -16,7 +16,7 @@
 // </copyright>
 //-------------------------------------------------------------------------------
 
-namespace Appccelerate.StateMachine
+namespace Appccelerate.StateMachine.Sync
 {
     using System;
     using Appccelerate.StateMachine.Machine.Events;
