+++ source/Appccelerate.StateMachine.Specs/Sync/Guards.cs
@@ -16,10 +16,9 @@
 // </copyright>
 //-------------------------------------------------------------------------------
 
-namespace Appccelerate.StateMachine
+namespace Appccelerate.StateMachine.Sync
 {
     using FluentAssertions;
-
     using Xbehave;
 
     public class Guards
