+++ source/Appccelerate.StateMachine.Specs/Sync/HierarchicalStateMachineInitialization.cs
@@ -16,7 +16,7 @@
 // </copyright>
 //-------------------------------------------------------------------------------
 
-namespace Appccelerate.StateMachine
+namespace Appccelerate.StateMachine.Sync
 {
     using FluentAssertions;
     using Xbehave;
