+++ source/Appccelerate.StateMachine.Specs/Sync/PassiveStateMachines.cs
@@ -1,5 +1,5 @@
 ﻿//-------------------------------------------------------------------------------
-// <copyright file="ActiveStateMachines.cs" company="Appccelerate">
+// <copyright file="PassiveStateMachines.cs" company="Appccelerate">
 //   Copyright (c) 2008-2017 Appccelerate
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,28 +16,23 @@
 // </copyright>
 //-------------------------------------------------------------------------------
 
-namespace Appccelerate.StateMachine
+namespace Appccelerate.StateMachine.Sync
 {
-    using System.Threading;
-
     using Appccelerate.StateMachine.Machine;
-
     using FakeItEasy;
-
     using FluentAssertions;
-
     using Xbehave;
 
-    public class ActiveStateMachines
+    public class PassiveStateMachines
     {
         [Scenario]
         public void DefaultStateMachineName(
-            ActiveStateMachine<string, int> machine,
+            PassiveStateMachine<string, int> machine,
             StateMachineNameReporter reporter)
         {
-            "establish an instantiated active state machine"._(() =>
+            "establish an instantiated passive state machine"._(() =>
             {
-                machine = new ActiveStateMachine<string, int>();
+                machine = new PassiveStateMachine<string, int>();
             });
 
             "establish a state machine reporter"._(() =>
@@ -50,19 +45,19 @@ namespace Appccelerate.StateMachine
 
             "it should use the type of the state machine as name for state machine"._(() =>
                 reporter.StateMachineName
-                    .Should().Be("Appccelerate.StateMachine.ActiveStateMachine<System.String,System.Int32>"));
+                    .Should().Be("Appccelerate.StateMachine.PassiveStateMachine<System.String,System.Int32>"));
         }
 
         [Scenario]
         public void CustomStateMachineName(
-            ActiveStateMachine<string, int> machine,
+            PassiveStateMachine<string, int> machine,
             StateMachineNameReporter reporter)
         {
             const string Name = "custom name";
 
-            "establish an instantiated active state machine with custom name"._(() =>
+            "establish an instantiated passive state machine with custom name"._(() =>
             {
-                machine = new ActiveStateMachine<string, int>(Name);
+                machine = new PassiveStateMachine<string, int>(Name);
             });
 
             "establish a state machine reporter"._(() =>
@@ -80,7 +75,7 @@ namespace Appccelerate.StateMachine
 
         [Scenario]
         public void CustomFactory(
-            ActiveStateMachine<string, int> machine,
+            PassiveStateMachine<string, int> machine,
             StandardFactory<string, int> factory)
         {
             "establish a custom factory"._(() =>
@@ -88,9 +83,9 @@ namespace Appccelerate.StateMachine
                 factory = A.Fake<StandardFactory<string, int>>();
             });
 
-            "when creating an active state machine"._(() =>
+            "when creating a passive state machine"._(() =>
             {
-                machine = new ActiveStateMachine<string, int>("_", factory);
+                machine = new PassiveStateMachine<string, int>("_", factory);
 
                 machine.In("initial").On(42).Goto("answer");
             });
@@ -101,21 +96,20 @@ namespace Appccelerate.StateMachine
 
         [Scenario]
         public void EventsQueueing(
-            IStateMachine<string, int> machine,
-            AutoResetEvent signal)
+            IStateMachine<string, int> machine)
         {
             const int FirstEvent = 0;
             const int SecondEvent = 1;
 
-            "establish an active state machine with transitions"._(() =>
-            {
-                signal = new AutoResetEvent(false);
+            bool arrived = false;
 
-                machine = new ActiveStateMachine<string, int>();
+            "establish a passive state machine with transitions"._(() =>
+            {
+                machine = new PassiveStateMachine<string, int>();
 
                 machine.In("A").On(FirstEvent).Goto("B");
                 machine.In("B").On(SecondEvent).Goto("C");
-                machine.In("C").ExecuteOnEntry(() => signal.Set());
+                machine.In("C").ExecuteOnEntry(() => arrived = true);
 
                 machine.Initialize("A");
             });
@@ -128,26 +122,25 @@ namespace Appccelerate.StateMachine
             });
 
             "it should queue event at the end"._(() =>
-                signal.WaitOne(1000).Should().BeTrue("state machine should arrive at destination state"));
+                arrived.Should().BeTrue("state machine should arrive at destination state"));
         }
 
         [Scenario]
         public void PriorityEventsQueueing(
-            IStateMachine<string, int> machine,
-            AutoResetEvent signal)
+            IStateMachine<string, int> machine)
         {
             const int FirstEvent = 0;
             const int SecondEvent = 1;
 
-            "establish an active state machine with transitions"._(() =>
-            {
-                signal = new AutoResetEvent(false);
+            bool arrived = false;
 
-                machine = new ActiveStateMachine<string, int>();
+            "establish a passive state machine with transitions"._(() =>
+            {
+                machine = new PassiveStateMachine<string, int>();
 
                 machine.In("A").On(SecondEvent).Goto("B");
                 machine.In("B").On(FirstEvent).Goto("C");
-                machine.In("C").ExecuteOnEntry(() => signal.Set());
+                machine.In("C").ExecuteOnEntry(() => arrived = true);
 
                 machine.Initialize("A");
             });
@@ -160,7 +153,7 @@ namespace Appccelerate.StateMachine
             });
 
             "it should queue event at the front"._(() =>
-                signal.WaitOne(1000).Should().BeTrue("state machine should arrive at destination state"));
+                arrived.Should().BeTrue("state machine should arrive at destination state"));
         }
     }
 }
